backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pdfplumber
python-docx
python-pptx
openpyxl
requests
